Fifth Checkpoint

+++ clear_database.py
@@ -9,6 +9,8 @@ from get_embedding_function import get_embedding_function
 # from langchain_community.vectorstores import Chroma
 from langchain_chroma import Chroma
 from langchain_community.document_loaders import PyPDFDirectoryLoader
+from logger_utils import setup_logger
+logger = setup_logger()
 
 
 CHROMA_PATH = "chroma"
@@ -19,9 +21,12 @@ def main():
 
 def clear_database(db):
     try:
-        print("✨ Clearing Database")
+        logger.info("✨ Clearing Database")
         ids_to_delete = db.get()["ids"]
-        print(ids_to_delete)        
+        logger.info(f'ids_to_delete: {ids_to_delete}')
+        if not ids_to_delete:
+            logger.info("No IDs to delete.")
+            return False        
         db.delete(ids=ids_to_delete)
         # db._client._system.stop()
         # SharedSystemClient._identifer_to_system.pop(vector_db._client._identifier, None)
@@ -30,6 +35,8 @@ def clear_database(db):
         #     shutil.rmtree(CHROMA_PATH)
         with open(AVAILABLE_FILES_PATH, "w") as file:
             file.write("")
+            logger.info(f"Cleared {AVAILABLE_FILES_PATH} file.")
+        logger.info("✨ Database cleared successfully.")
         return True
     except Exception as e:
         print(e)
